+++ configurio/configs/yaml.py
@@ -1,11 +1,6 @@
-from logging import getLogger
-
 import yaml
 
 from ..core import BaseConfiguration
-
-
-logger = getLogger(__name__)
 
 
 class YAML_Configuration(BaseConfiguration):