+++ configurio/tests/core.py
@@ -49,12 +49,20 @@ class BaseConfigTest():
         self.config = BaseConfiguration()
 
     def test_delete_file(self):
-        """ Delete local configuration file """
-        self.config.delete_file()
+        """
+        Delete local configuration file and check
+        is the functions output works correctly
+        """
+        expect_success = self.config.delete_file()
+        expect_fail = not self.config.delete_file()
+
+        assert expect_success and expect_fail
 
     def test_create_file(self):
         """ Create local configuration file """
         self.config.create_file()
+
+        assert self.config.is_file_exist()
 
     def test_file_read(self):
         # TODO
@@ -65,10 +73,6 @@ class BaseConfigTest():
         pytest.skip("Not finished")
 
     def test_modify_file(self):
-        """
-        Modify configuration object attribute value and commit changes to local file.
-        Then check if the content of file is equal to the content of modified attribute.
-        """
         person_name = "Jeff"
         self.config['Person_1']['name'] = person_name
         self.config.commit()
@@ -77,35 +81,29 @@ class BaseConfigTest():
 
         assert read_dict["Person_1"]["name"] == person_name
 
-    def test_reset_file_to_defaults(self):
-        """ Reset local file to values from bound default configuration """
-        self.config.reset_file_to_defaults()
-
-        assert self.config.read_file_as_dict()["Person_1"]["name"] == default_configuration_dict["Person_1"]["name"]
-
-    def test_clear_object(self):
-        """ Clear all object's attributes """
+    def test_clear_dict(self):
+        """ Clear the `dictionary` attribute """
         self.config.clear()
 
         assert len(self.config) == 0
 
-    def test_refresh_object(self):
-        """ Read all attributes from local file """
+    def test_refresh_dict(self):
+        """ Refresh `dictionary` from local file """
         self.config.refresh()
 
         assert len(vars(self.config)) > 0
 
-    def test_reset_object_to_file(self):
-        """ Reset all object's attributes to values from bound configuration file """
+    def test_reset_dict_to_file(self):
+        """ Reset `dictionary` to values from bound configuration file """
         self.config['tobereset'] = False
         self.config.reset_to_file()
 
         assert self.config.get('tobereset', True)
 
-    def test_access_dashed_attribute_like_dict(self):
+    def test_access_dashed_attribute(self):
         assert self.config["dash-split"]["wow-man"] == default_configuration_dict["dash-split"]["wow-man"]
 
-    def test_add_dashed_attribute_like_dict(self):
+    def test_add_dashed_attribute(self):
         self.config["dash-split"]["second-dashed"] = 515
         self.config.commit()
 
@@ -114,8 +112,8 @@ class BaseConfigTest():
 
     def test_add_one_attribute(self):
         """
-        Add one new attribute to object, commit changes
-        and check assert the local file to object
+        Add one new attribute to `dictionary`, commit changes
+        and assert the local file to `dictionary`
         """
         key_name = "Masvingo"
 
@@ -126,8 +124,8 @@ class BaseConfigTest():
 
     def test_add_nested_dicts(self):
         """
-        Add nested dictionary to object, commit changes and assert
-        local file values to object
+        Add nested dictionary to `dictionary`, commit changes and assert
+        local file values to `dictionary`
         """
         dct = {
             "sugar": {
@@ -142,7 +140,7 @@ class BaseConfigTest():
 
     def test_add_multiple_dicts_with_update(self):
         """
-        Add multiple dicts to object, commit changes
+        Add multiple dicts to `dictionary`, commit changes
         and assert the result with local file
         """
         dct = {
@@ -161,3 +159,10 @@ class BaseConfigTest():
 
         assert self.config['users']['jerre_1829']['status'] == file_dict["jerre_1829"]["status"] and \
                self.config['users']['wolfrm4882']['status'] == file_dict["wolfrm4882"]["status"]
+
+    def test_modify_default_dict(self):
+        """Modify the default dictionary, reset `dictionary` to defaults and assert values"""
+        self.config.dictionary_default.update({"new_values": [256, 45, 154]})
+        self.config.reset_to_defaults()
+
+        assert self.config.get('new_values', None) == self.config.dictionary_default['new_values']